+++ src/Snippet/Manager.php
@@ -4,14 +4,8 @@ declare(strict_types=1);
 
 namespace Bolt\Snippet;
 
-use Bolt\Widget\BaseWidget;
-use Bolt\Widget\BoltHeaderWidget;
-use Bolt\Widget\CanonicalLinkWidget;
-use Bolt\Widget\NewsWidget;
-use Bolt\Widget\WeatherWidget;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
-use Symfony\Component\HttpFoundation\Response;
 use Tightenco\Collect\Support\Collection;
 
 class Manager
@@ -22,33 +16,15 @@ class Manager
     /** @var Request */
     private $request;
 
-    /** @var QueueProcessor */
-    private $queueProcessor;
-
-    public function __construct(RequestStack $requestStack, QueueProcessor $queueProcessor)
+    public function __construct(RequestStack $requestStack)
     {
         $this->queue = new Collection([]);
         $this->request = $requestStack->getCurrentRequest();
-        $this->queueProcessor = $queueProcessor;
     }
 
-    /**
-     * @param BaseWidget|string|callable $callback
-     */
-    public function registerSnippet(
-        $callback,
-        string $target = Target::NOWHERE,
-        string $zone = Zone::FRONTEND,
-        string $name = 'nameless snippet',
-        int $priority = 100
-    ): void {
-        $this->queue->push([
-            'priority' => $priority,
-            'target' => $target,
-            'name' => $name,
-            'zone' => $zone,
-            'callback' => $callback,
-        ]);
+    public function registerSnippet($name, $callback): void
+    {
+        dump('register Snippet ' . $name);
     }
 
     public function getSnippet($name): void
@@ -70,8 +46,10 @@ class Manager
 
         $output = '';
 
-        foreach ($widgets as $widget) {
-            $output .= $this->invoke($twig, $widget['callback']);
+        if ($widgets) {
+            foreach ($widgets as $widget) {
+                $output .= $this->invoke($twig, $widget['callback']);
+            }
         }
 
         return $output;
@@ -97,7 +75,6 @@ class Manager
             'priority' => $widget->getPriority(),
             'target' => $widget->getTarget(),
             'name' => $widget->getName(),
-            'zone' => $widget->getZone(),
             'callback' => $widget,
         ]);
     }
@@ -107,17 +84,10 @@ class Manager
         $this->registerWidget(new WeatherWidget());
         $this->registerWidget(new NewsWidget());
 
-        $this->registerSnippet('<meta name="generator" content="Bolt">', Target::END_OF_HEAD);
+        $this->registerSnippet(ZONE::FRONTEND, Target::END_OF_HEAD, '<meta name="generator" content="Bolt">');
         $this->registerWidget(new CanonicalLinkWidget());
-        $this->registerWidget(new BoltHeaderWidget());
 
         // @todo Determine if a favicon is something we want in 2019, by default
         // $this->registerWidget(new FaviconWidget());
-    }
-
-    public function processQueue(Response $response): void
-    {
-        $zone = Zone::get($this->request);
-        $this->queueProcessor->process($response, $this->queue, $zone);
     }
 }
