--- src/Snippets/BaseSnippet.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Bolt\Snippets;
-
-class BaseSnippet
-{
-}
