+++ src/Widget/CanonicalLinkWidget.php
@@ -2,16 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Bolt\Widget;
-
-use Bolt\Snippet\Target;
-use Bolt\Snippet\Zone;
+namespace Bolt\Snippet;
 
 class CanonicalLinkWidget extends BaseWidget
 {
     protected $name = 'Weather Widget';
     protected $type = 'widget';
-    protected $target = Target::NOWHERE;
-    protected $zone = Zone::NOWHERE;
+    protected $target = Target::END_OF_HEAD;
+    protected $zone = Zone::FRONTEND;
     protected $priority = 200;
 }
